Kodlama.io

+++ Entities/DTOs/CarDetailDto.cs
@@ -11,6 +11,7 @@ namespace Entities.DTOs
         public string CarName { get; set; }
         public string BrandName { get; set; }
         public string ColorName { get; set; }
-        public int DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
     }
 }
+
